# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "AGENT STATS",
    "description": "AGENT STATS STATUSES AND CHANNEL DATA"
  },
  "servers": [
    {
      "url": "http://localhost:3000/"
    },
    {
      "url": "https://localhost:3000/"
    }
  ],
...
  "paths": {
...
    "/job/upload-data": {
      "post": {
        "tags": [
          "Jobs"
        ],
        "description": "get jobs today",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string"
                  },
                  "pause": {
                    "type": "number"
                  },
                  "complete": {
                    "type": "number"
                  },
                  "startTime": {
                    "type": "string"
                  },
                  "endTime": {
                    "type": "string"
                  },
                  "active": {
                    "type": "number"
                  },
                  "queue": {
                    "type": "string"
                  },
                  "concurrency": {
                    "type": "number"
                  },
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/job/": {
      "post": {
        "tags": [
          "Job"
        ],
        "description": "create new out dial jobs.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "obj",
            "in": "body",
            "description": "User information.",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CreateJob"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      },
      "get": {
        "tags": [
          "Job"
        ],
        "description": "all sip peers.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "all sip peers.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SipPeers"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/SipPeers"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/job/{id}": {
      "put": {
        "tags": [
          "Job"
        ],
        "description": "update outbound job by id.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string"
                  },
                  "pause": {
                    "type": "number"
                  },
                  "complete": {
                    "type": "number"
                  },
                  "startTime": {
                    "type": "string"
                  },
                  "endTime": {
                    "type": "string"
                  },
                  "active": {
                    "type": "number"
                  },
                  "queue": {
                    "type": "string"
                  },
                  "concurrency": {
                    "type": "number"
                  },
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Job"
        ],
        "description": "update outbound job by id.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "obj",
            "in": "body",
            "description": "User information.",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PatchJob"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Job"
        ],
        "description": "all sip peers.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "all sip peers.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SipPeers"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/SipPeers"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      },
      "get": {
        "tags": [
          "Job"
        ],
        "description": "all sip peers.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "all sip peers.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SipPeers"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/SipPeers"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/job/types": {
      "get": {
        "tags": [
          "Job"
        ],
        "description": "all sip peers.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "all sip peers.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SipPeers"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/SipPeers"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/job/cdrs": {
      "post": {
        "tags": [
          "Job"
        ],
        "description": "create new out dial jobs.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "obj",
            "in": "body",
            "description": "User information.",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CreateJob"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/job/cdr-record/{id}": {
      "delete": {
        "tags": [
          "Job"
        ],
        "description": "all sip peers.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "all sip peers.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SipPeers"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/SipPeers"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/job/job-cdrs/{jobId}": {
      "delete": {
        "tags": [
          "Job"
        ],
        "description": "all sip peers.",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "all sip peers.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SipPeers"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/SipPeers"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/job/inject": {
      "post": {
        "tags": [
          "Job"
        ],
        "description": "all sip peers.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "all sip peers.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Inject"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Inject"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/job/status/{id}": {
      "get": {
        "tags": [
          "Job"
        ],
        "description": "all sip peers.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "all sip peers.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Inject"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Inject"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/job/phone-call/{jobId}/{id}": {
      "delete": {
        "tags": [
          "Job"
        ],
        "description": "all sip peers.",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "all sip peers.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SipPeers"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/SipPeers"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/job/inactive-job/{id}": {
      "delete": {
        "tags": [
          "Job"
        ],
        "description": "all sip peers.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "all sip peers.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SipPeers"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/SipPeers"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "CreateJob": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "test"
          },
          "type": {
            "type": "string",
            "example": "predictive"
          },
          "pause": {
            "type": "number",
            "example": 0
          },
          "complete": {
            "type": "number",
            "example": 0
          },
          "startTime": {
            "type": "string",
            "example": "2022-08-10 20:00:00"
          },
          "endTime": {
            "type": "string",
            "example": "2022-08-10 23:00:00"
          },
          "jobData": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "number": {
                  "type": "string",
                  "example": "231203"
                },
                "mask": {
                  "type": "string",
                  "example": "123456"
                },
                "context": {
                  "type": "string",
                  "example": "default"
                },
                "trunk": {
                  "type": "string",
                  "example": "outgoing"
                }
              }
            }
          }
        },
        "xml": {
          "name": "CreateJob"
        }
      },
...
      "SipPeers": {
        "type": "object",
        "properties": {
          "response": {
            "type": "object",
            "properties": {
              "success": {
                "type": "boolean",
                "example": true
              },
              "data": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "event": {
                      "type": "string",
                      "example": "PeerEntry"
                    },
                    "actionid": {
                      "type": "string",
                      "example": "1660286907216"
                    },
                    "channeltype": {
                      "type": "string",
                      "example": "SIP"
                    },
                    "objectname": {
                      "type": "string",
                      "example": "salman"
                    },
                    "chanobjecttype": {
                      "type": "string",
                      "example": "peer"
                    },
                    "ipaddress": {
                      "type": "string",
                      "example": "-none-"
                    },
                    "ipport": {
                      "type": "string",
                      "example": "0"
                    },
                    "dynamic": {
                      "type": "string",
                      "example": "yes"
                    },
                    "autoforcerport": {
                      "type": "string",
                      "example": "no"
                    },
                    "forcerport": {
                      "type": "string",
                      "example": "no"
                    },
                    "autocomedia": {
                      "type": "string",
                      "example": "no"
                    },
                    "comedia": {
                      "type": "string",
                      "example": "no"
                    },
                    "videosupport": {
                      "type": "string",
                      "example": "no"
                    },
                    "textsupport": {
                      "type": "string",
                      "example": "no"
                    },
                    "acl": {
                      "type": "string",
                      "example": "no"
                    },
                    "status": {
                      "type": "string",
                      "example": "UNKNOWN"
                    },
                    "realtimedevice": {
                      "type": "string",
                      "example": "no"
                    },
                    "description": {
                      "type": "string",
                      "example": ""
                    },
                    "accountcode": {
                      "type": "string",
                      "example": ""
                    }
                  }
                }
              }
            }
          }
        },
        "xml": {
          "name": "SipPeers"
        }
      },
...
      "PatchJob": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "test"
          },
          "type": {
            "type": "string",
            "example": "predictive"
          },
          "pause": {
            "type": "number",
            "example": 0
          },
          "complete": {
            "type": "number",
            "example": 0
          },
          "startTime": {
            "type": "string",
            "example": "2022-08-10 20:00:00"
          },
          "endTime": {
            "type": "string",
            "example": "2022-08-10 23:00:00"
          }
        },
        "xml": {
          "name": "PatchJob"
        }
      },
...
    },
    "securitySchemes": {
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "any description..."
      }
    }
  }
}